WIP: Dead end?

+++ src/main/java/assertionbit/trainapi/messages/TicketGroupRequst.java
@@ -3,6 +3,6 @@ package assertionbit.trainapi.messages;
 import java.util.ArrayList;
 
 public class TicketGroupRequst {
-    public ArrayList<Integer> wagonId;
-    public ArrayList<Integer> sitId;
+    public ArrayList<Long> wagonId;
+    public ArrayList<Long> sitId;
 }
